refactor: split userhome.go into per-page files
